requirements.txt
# python=3.8.3
scikit-image==0.18.0
tqdm==4.57.0
numpy==1.21.1
pillow==8.1.1
fastapi
uvicorn[standard]